+++ source/did-ta-server/src/main/java/org/omnione/did/tas/v1/common/service/MockStorageServiceImpl.java
@@ -5,13 +5,15 @@ import org.omnione.did.data.model.did.DidDocument;
 import org.omnione.did.data.model.did.InvokedDidDoc;
 import org.omnione.did.data.model.enums.did.DidDocStatus;
 import org.omnione.did.data.model.enums.vc.RoleType;
+import org.omnione.did.data.model.enums.vc.VcStatus;
+import org.omnione.did.data.model.schema.VcSchema;
 import org.omnione.did.data.model.vc.VcMeta;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
-@Profile("sample") // ✅ sample 프로파일일 때만 등록
+@Profile("sample")
 public class MockStorageServiceImpl implements StorageService {
     @Override
     public void registerDidDoc(InvokedDidDoc didDoc, RoleType roleType) {
@@ -19,8 +21,7 @@ public class MockStorageServiceImpl implements StorageService {
     }
 
     @Override
-    public DidDocument updateDidDocStatus(String did, DidDocStatus didDocStatus) {
-        return null;
+    public void updateDidDocStatus(String did, DidDocStatus didDocStatus) {
     }
 
     @Override
@@ -32,9 +33,22 @@ public class MockStorageServiceImpl implements StorageService {
     public void registerVcMeta(VcMeta vcMeta) {
 
     }
+    @Override
+    public void updateVcMeta(String vcId, VcStatus vcStatus) {
+    }
 
     @Override
     public VcMeta findVcMeta(String vcId) {
+        return null;
+    }
+
+    @Override
+    public void registerVcSchema(VcSchema vcSchema, String did) {
+
+    }
+
+    @Override
+    public VcSchema getVcSchema(String vcSchemaId) {
         return null;
     }
 }